+++ NEWS.md
@@ -4,6 +4,7 @@
 
 * Return NA for documents that do not have known features 
 * Drop document variables to avoid slowdown and warnings
+* Corrrect Japanese seed words for DE, MG and EC 
 
 ## New data